fixed cursed types in render.cpp
2 bit int for one byte color value O_O

+++ source/nds/render.cpp
@@ -55,7 +55,7 @@ void Render::renderSprites() {
 void Render::renderVisibleVariables() {
 }
 
-void Render::drawBox(int w, int h, int x, int y, int colorR, int colorG, int colorB, int colorA) {
+void Render::drawBox(int w, int h, int x, int y, uint8_t colorR, uint8_t colorG, uint8_t colorB, uint8_t colorA) {
 gl2d.box(x,y,w,h,colorR,colorB,colorG,colorA);
 }
